Rename or remove files not named `*_test.py`; fix Cypress fixtures
